+++ src/util/markdown-plugins.ts
@@ -5,7 +5,7 @@ type NodeNodeMap = (node: TNodeValue) => TNodeValue;
 type NodeStringMap = (node: TNodeValue) => string;
 import { regexReplace } from './regexReplace';
 import { wikilinksToHypertextLinks } from './wikilinks';
-import {paragraphRegex} from './paragraphRegex'
+import {paragraphRegex} from './paragraphIdCreator'
 
 const nodeReplacer: (
     node: TNodeValue,
@@ -68,11 +68,8 @@ export const fixObsidianDashes: NodeNodeMap = node =>
             x => '–'
         ),
     });
-export const paragraphLinks: NodeNodeMap = node =>
-    Object.assign(node, {
-        type: 'html',
-        value: paragraphRegex(node.value)
-    });
+export const paragraphLinks = (node: Element) =>
+    {paragraphRegex(node)};
 export const obsidianWikilinks: (files: string[]) => NodeNodeMap =
     files => node =>
         Object.assign(node, {
